Move entity tabs to bottom tab bar
